Update utils.py
Add a function getFigures()
If the target is found at the end of the expansion, create figures of the
reaction routes. Each figure contains all generations of one route. Also 
prints reaction operators, reactants, and products. Figures are displayed,
also saved in the working folder. Based on the example code for using 
RxnTrackerDepthFirst(rxn_lib).
Currently need to have the font file 'OpenSans-Regular.ttf' in the working 
folder.